+++ CHANGELOG.markdown
@@ -1,6 +1,9 @@
 Changelog
 =========
 
+0.3
+---
+* Added `click` proxy. Ordinary built-in `click` can now be used to establish long click event simply by supplying a duration.
 0.2
 ---
 * `longclick` method now accepts an optional duration argument which overrides the default duration.